openrc-run: exit if a required file/dir is missing
Commit 29f7e335927d4c1fb151124de8cdf01fb87723fa modified the required
path searches to read the paths from a pipe.  This causes the while loop
to execute in a subshell.  If a required path was missing the `exit 1`
was only exiting the subshell, and openrc-run continued to run.  Using
input redirection rather than a pipe keeps the while loop in the main
process allowing openrc-run to exit if a path is missing.

Signed-off-by: Andrew Gregory <andrew.gregory.8@gmail.com>